✍️Rename($Lotto): 모듈 경로 이동
